Added test to verify if the MatrixProduct returns the correct value when n*n is higher than int. Changed the return type, so the test passes.

+++ src/primetables/IMatrixComputationBehaviour.java
@@ -6,6 +6,6 @@ package primetables;
  */
 public interface IMatrixComputationBehaviour {
 
-    public int[][] calculate(int[] n);
+    public long[][] calculate(int[] n);
     
 }
