+++ src/primetables/MatrixProduct.java
@@ -7,14 +7,14 @@ package primetables;
 public class MatrixProduct implements IMatrixComputationBehaviour {
 
     @Override
-    public int[][] calculate(int[] factorsArray) {
+    public long[][] calculate(int[] factorsArray) {
         if(factorsArray==null){
             return null;}
         int size = factorsArray.length;
-        int [][] matrixProduct = new int[size][size];
+        long [][] matrixProduct = new long[size][size];
         for(int i =0;i<size;i++){
             for (int j=0;j<size;j++){
-                matrixProduct[i][j]=factorsArray[i]*factorsArray[j];
+                matrixProduct[i][j]=(long)factorsArray[i]*(long)factorsArray[j];
             }
         }
         return matrixProduct;
